chore: add json endpoint
