+++ build.gradle
@@ -25,6 +25,7 @@ dependencies {
 	implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
 	compileOnly 'org.projectlombok:lombok:1.18.34'
 	annotationProcessor 'org.projectlombok:lombok:1.18.34'
+	implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
 }
 
 tasks.named('test') {
