+++ src/main/kotlin/hu/telekom/tkriszti/flightInfo/controller/AppControllerJson.kt
@@ -3,21 +3,19 @@ package hu.telekom.tkriszti.flightInfo.controller
 import hu.telekom.tkriszti.flightInfo.dto.ResultDTO
 import hu.telekom.tkriszti.flightInfo.service.PilotInfoService
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.stereotype.Controller
-import org.springframework.ui.Model
+import org.springframework.http.MediaType
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestParam
+import org.springframework.web.bind.annotation.RestController
 import java.sql.SQLException
 
-@Controller
-class AppController
+@RestController
+class AppControllerJson
     (@Autowired val service: PilotInfoService) {
 
-        @GetMapping("/pilot")
+        @GetMapping("/pilot/json", produces = [MediaType.APPLICATION_JSON_VALUE])
         @Throws(SQLException::class)
-        fun showPilotData(model: Model, @RequestParam("name") name: String): String {
-            val pilotDtos: Set<ResultDTO?>? = service.getPilotData(name)
-            model.addAttribute("pilot", pilotDtos)
-            return "pilot.html"
+        fun getPilotDataJson(@RequestParam("name") name: String): Set<ResultDTO?>? {
+            return service.getPilotData(name)
         }
     }
